mmu: write eram to sav file on display refresh
Previously, we only wrote the eram to disk when the program exists. The
purpose was that we don't want to write it on every eram change, since
it will be accessed one byte at a time.

A nice solution to this problem is to only write out the eram when we
refresh the display. This happens about every 1/60 seconds, so when
saving a game we should only end up writing the sav file once or twice
while the game is updating it.