gbemuc: add protura backend
This adds a backend driver that makes use of the framebuffer provided by
Protura to render. It does not support sound and only supports graphics.

Keyboard input is grabbed via the /dev/keyboard event stream. We also
use a bunch of `ioctl()`s to setup the screen, keyboard, and
framebuffer.