Protura: fix usleep() delay logic to be more accurate
The previous logic was wrong because it assumed usleep() sleeps for the
exact amount of time we feed it, when that's actually incorrect. Protura
previously was incorrectly rounding this value down, resulting in the
game playing slightly too fast. However, now that Protura has been fixed
it actually runs too slow as Protura rounds the value up and we sleep
for a fair amount longer time.

The fix is to check how long the usleep() actually took and keep track
of the "difference" between how long we should have been running for,
and how long we actually have been running for. The difference is then
the amount we sleep by, but we keep a rolling count and simply subtract
what the time is after the sleep. For sleeps that go long, the
difference will go negative and will result in a shorter sleep next
time.

The result is timing that is extremely close to the 59.7 that we're
going for.